requirements.txt
click==8.1.7
jsonschema==4.20.0
pydantic==2.5.0
requests==2.31.0
python-dateutil==2.8.2
openai==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
google-generativeai==0.3.2